refactor(token-indicators): extract normalize.js's per-kind logic into kinds/ modules
